+++ ark-common/src/main/java/au/org/theark/core/dao/IArkSettingDao.java
@@ -18,7 +18,7 @@ public interface IArkSettingDao {
 
     public int getSettingsCount(Setting setting);
 
-    public List<Setting> searchPageableSettings(Setting setting, long first, long count);
+    public List<Setting> searchPageableSettings(Setting setting, int first, int count);
 
     public Setting getUserSpecificSetting(ArkUser arkUser, String propertyName);
