+++ ark-common/src/main/java/au/org/theark/core/service/IArkCommonService.java
@@ -27,9 +27,12 @@ import java.util.List;
 import java.util.Set;
 
 import au.org.theark.core.dao.IArkAuthorisation;
+import au.org.theark.core.model.study.entity.*;
 import org.apache.velocity.exception.VelocityException;
 import org.apache.wicket.markup.html.WebComponent;
 import org.apache.wicket.markup.html.form.upload.FileUpload;
+import org.hibernate.Criteria;
+import org.hibernate.criterion.Restrictions;
 import org.springframework.mail.MailSendException;
 import org.springframework.mail.SimpleMailMessage;
 
@@ -56,6 +59,7 @@ import au.org.theark.core.model.lims.entity.BioCollectionUidToken;
 import au.org.theark.core.model.lims.entity.BiospecimenUidPadChar;
 import au.org.theark.core.model.lims.entity.BiospecimenUidTemplate;
 import au.org.theark.core.model.lims.entity.BiospecimenUidToken;
+import au.org.theark.core.model.pheno.entity.PhenoDataSetField;
 import au.org.theark.core.model.pheno.entity.PhenoDataSetFieldDisplay;
 import au.org.theark.core.model.report.entity.BiocollectionField;
 import au.org.theark.core.model.report.entity.BiospecimenField;
@@ -77,6 +81,7 @@ import au.org.theark.core.model.study.entity.ArkRolePolicyTemplate;
 import au.org.theark.core.model.study.entity.ArkUser;
 import au.org.theark.core.model.study.entity.ArkUserRole;
 import au.org.theark.core.model.study.entity.AuditHistory;
+import au.org.theark.core.model.study.entity.Consent;
 import au.org.theark.core.model.study.entity.ConsentAnswer;
 import au.org.theark.core.model.study.entity.ConsentOption;
 import au.org.theark.core.model.study.entity.ConsentStatus;
@@ -491,7 +496,7 @@ public interface IArkCommonService<T> {
 	 * 
 	 * @return Collection of SubjectVO
 	 */
-	public List<SubjectVO> searchPageableSubjects(SubjectVO subjectVoCriteria, long first, long count);
+	public List<SubjectVO> searchPageableSubjects(SubjectVO subjectVoCriteria, int first, int count);
 
 	/**
 	 * A generic interface that will return count of the subjects in the study
@@ -552,7 +557,7 @@ public interface IArkCommonService<T> {
 	 *            - number of items to return (for that page)
 	 * @return Collection of Custom Fields
 	 */
-	public List<CustomField> searchPageableCustomFields(CustomField customFieldCriteria, long first, long count);
+	public List<CustomField> searchPageableCustomFields(CustomField customFieldCriteria, int first, int count);
 
 	/**
 	 * A less generic interface that will return a list CustomFields specified
@@ -907,13 +912,13 @@ public interface IArkCommonService<T> {
 
 	public int getPipelineCount(Pipeline p);
 
-	public List<Pipeline> searchPageablePipelines(Pipeline object, long first, long count);
+	public List<Pipeline> searchPageablePipelines(Pipeline object, int first, int count);
 
 	public long getPipelineCount(Study study);
 
 	public int getProcessCount(Process object);
 
-	public List<Process> searchPageableProcesses(Process object, long first, long count);
+	public List<Process> searchPageableProcesses(Process object, int first, int count);
 
 	public void createProcess(Process modelObject);
 
@@ -1058,7 +1063,7 @@ public interface IArkCommonService<T> {
 	 * @param count
 	 * @return
 	 */
-	public List<CustomFieldCategory> searchPageableCustomFieldCategories(CustomFieldCategory customFieldCategoryCriteria, long first, long count);
+	public List<CustomFieldCategory> searchPageableCustomFieldCategories(CustomFieldCategory customFieldCategoryCriteria, int first, int count);
 	
 	/**
 	 * Category list By custom field Type.
